# perf(tspath): void []byte allocation in ToFileNameLowerCase