@camc314 2025-08-13
@jakebailey up to you if this is merged, I thought this had a perf benifit, but i explored the compiled code in godbolt, and it appears exactly the same machine code is generated .

Regardless, i think this is slightly more explicit behaviour.