@jakebailey 2025-08-13
Yeah, this kind of conversion I am not surprised does not affect the compile. I agree it's clearer.